long for embedding

+++ zoo/src/main/java/com/intel/analytics/zoo/pipeline/api/net/PytorchModel.java
@@ -93,7 +93,7 @@ class PytorchModel {
     static native long loadLossNative(String lossPath);
 
     static native JTensor[] modelForwardNative(
-            long nativeRef, boolean isTraining, float[][] storage, int[] offset, int[][] shape);
+            long nativeRef, boolean isTraining, long[][] storage, int[] offset, int[][] shape);
 
     static native JTensor[] modelBackwardNative(
             long nativeRef, float[][] storage, int[] offset, int[][] shape);
